requirements.txt
gym==0.20.0
tensorflow==2.7.1
tf-agents==0.11.0
pillow==8.1.0
pyglet==1.5.0
scipy==1.5.4
h5py==3.1.0
# transient deps
tensorflow-probability==0.15.0
# https://github.com/protocolbuffers/protobuf/issues/10051
protobuf==3.20.1
